Add files via upload
IPython notebook that is a compilation of the various stuff I've been working on. I wrote functions for each of the actions called from javascript above, the residual/strain function is not actually done yet, but the process for finding the upper limit for the source region is working properly now (once residual/strain equation is working properly it'll give us a detection significance).