Added eccentricity into residual calculation
Included the g(n, e) function from Peters & Matthews 1963.
Following Taylor et al 2015, we calculate the residual from the harmonic that contributes the most, which is n = 2 for 0 < ecc < 0.4, and n = 1 for ecc < 0.4. The resulting residual is then multiplied by the square root of the g(n, e) function for the appropriate values of n and e.